Refactor morphology module to break dependency cycles
Extracted core types (MorphAnalysis, PartOfSpeech, Case, etc.) into `src/morphology/models.rs`.
Updated submodules to import from `models` instead of `super`.
This creates a clean DAG dependency structure: mod -> lexicon -> models.
