+++ src/morphology/declension.rs
@@ -7,8 +7,8 @@
 
 use std::borrow::Cow;
 
-use super::{Case, Gender, MorphAnalysis, Number, PartOfSpeech};
 use crate::morphology::matcher::match_suffix;
+use crate::morphology::models::{Case, Gender, MorphAnalysis, Number, PartOfSpeech};
 
 /// Declension pattern
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
